refactor(extensions): update `DateTime` references to use `Date`
