+++ app/src/main/java/com/tunaikumobile/androidextension/MainActivity.kt
@@ -6,7 +6,6 @@ import com.tunaikumobile.androidextension.databinding.ActivityMainBinding
 import com.tunaikumobile.extensions.convertToString
 import com.tunaikumobile.extensions.isWeekend
 import com.tunaikumobile.extensions.plusOrMinusYears
-import org.joda.time.DateTime
 import java.util.Date
 
 class MainActivity : AppCompatActivity() {
@@ -19,11 +18,10 @@ class MainActivity : AppCompatActivity() {
         val view = binding.root
 
         val date = Date(System.currentTimeMillis())
-        val dateTime = DateTime(date)
         binding.tvToday.text = date.convertToString()
         binding.tvDateFormat.text = date.convertToString("dd/MMMM/yyyy")
         binding.tvPlusYear.text = date.plusOrMinusYears(1)
-        binding.tvWeekend.text = "Is today a holiday? ${dateTime.isWeekend()}"
+        binding.tvWeekend.text = "Is today a holiday? ${date.isWeekend()}"
         setContentView(view)
     }
 }
