+++ extensions/src/androidTest/java/com/tunaikumobile/extensions/DateExtensionsTest.kt
@@ -2,13 +2,22 @@ package com.tunaikumobile.extensions
 
 
 import org.junit.Assert
+import org.junit.Before
 import org.junit.Test
+import java.util.*
 
 /**
  * @author Kasmadi
  * @date 15/11/23
  */
 class DateExtensionsTest {
+
+    private lateinit var date:Date
+
+    @Before
+    fun setUp() {
+        date = "2023-11-14".convertToDate()
+    }
 
     @Test
     fun getDiffDaysTest() {
@@ -79,5 +88,28 @@ class DateExtensionsTest {
         val date = "2023-11-14".convertToDate()
         Assert.assertTrue(date.changeDateFormat("dd/MM/yyyy") == "14/11/2023".convertToDate("dd/MM/yyyy"))
 
+    }
+
+    @Test
+    fun getShortMontTest() {
+        Assert.assertTrue(date.getShortMonth() == "Nov")
+    }
+
+    @Test
+    fun getDayAsTextTest() {
+        Assert.assertTrue(date.getDayAsText() == "Selasa")
+    }
+
+    @Test
+    fun isWeekendReturnFalse() {
+        Assert.assertFalse(date.isWeekend())
+    }
+
+    @Test
+    fun isWeekendReturnTrue() {
+        val date1 = "2023-11-11".convertToDate()
+        val date2 = "2023-11-12".convertToDate()
+        Assert.assertTrue(date1.isWeekend())
+        Assert.assertTrue(date2.isWeekend())
     }
 }
